feat: limit open ended vectors in chat (#6467)
Description
---
- Limited open-ended vectors in chat to guard against malicious network
messages.
- Fixed instability issues with chat FFI cucumber tests.

Motivation and Context
---
This is a _defense-in-depth_ exercise.

How Has This Been Tested?
---
- Existing unit tests pass.
- Added new unit tests to test all limits

What process can a PR reviewer use to test or verify this change?
---
Code review

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->