fix: contact service tari address support (#6478)
Description
---
Contact service still assumes that the node_id and tari address are the
same. This changes contact service to check the tari address.

Motivation and Context
---
This allows a user to update the address of a contact without having to
delete the contact. It will look at the address given and update the
address accordingly.

How Has This Been Tested?
---
Manual