Add custom noise prologue
Added a custom noise prologue to the noise protocol to be hashed in to the handshake hash value.
This enables an application to specify a unique identifier like the genesis block hash as the
noise prologue, effectively stopping any communication handshake to succeed with nodes that is
not on the same genesis block.