fix(comms): always send disconnect notification to conn manager (#6511)
Description
---
fix(comms): always send disconnect notification to conn manager
fix(comms): take the connection state into account when determining
connectedness for tie-breaking

Motivation and Context
---
Previously, if Yamux detected that we were disconnected, the
connectivity manager `Disconnected` notification would not be sent. A
race condition with disconnects on a connection and the remote
disconnecting happen at almost the same time. If the latter happens
first, then the notification of the disconnect would not be sent to the
connection manager. When a new connection is attempted, the connection
manager would think that a previous connection exists.

Ref #6512 
Ref https://github.com/tari-project/tari/issues/6513

How Has This Been Tested?
---
Not explicitly tested, difficult to reproduce.

What process can a PR reviewer use to test or verify this change?
---

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->