feat: add network wire byte (#6518)
Description
---
Added network wire byte. This is now independent of the network as byte.

Note: This method in favour of #6502.

Motivation and Context
---
Base nodes and wallets maintained numerous connections from which they
could not sync or obtain useful information. This caused base nodes to
try and follow other base nodes that were on a higher proof of work but
on a different genesis block, and caused wallets to query base nodes
that supplied information from a different blockchain. This byte can now
be changed every time the genesis block changes.

How Has This Been Tested?
---
Passed all unit tests.

What process can a PR reviewer use to test or verify this change?
---

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [X] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->