feat: add configurable exclude dialer addresses for universe (#6543)
Description
---
Added user-configurable communication node addresses that should never
be dialled. In Tari Universe the base node and wallet is configured in a
listen-only TCP mode with a fake public address and when their addresses
are propagated throughout the network, other peers should not try to
dial them.

Motivation and Context
---
The previous implementation was hard-coded and did not work properly.

How Has This Been Tested?
---
Added unit tests.
System-level testing.

What process can a PR reviewer use to test or verify this change?
---
Code review.

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->