feat: add generic config list type and use throughout all config settings (#6567)
Description
---
Added a generic config list type for use with config files and applied
this to config arrays of type 'Vec<T>'. With this PR, all array-type
config options can be omitted in the config file and/or passed in via
the command line as a config property override with `-p <override>'.

Motivation and Context
---
Using config options where arrays have been used was not flexible or
consistent throughout the config implementation.

How Has This Been Tested?
---
Added unit tests.
System-level testing.

What process can a PR reviewer use to test or verify this change?
---
Code review.
System-level testing.

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->