feat: add wallet ffi backup peers (#6591)
Description
---
Added wallet FFI backup peers (all seed peers) when the base node is
set.

Motivation and Context
---
This adds resilience to the base node connection.

How Has This Been Tested?
---
The scheme works for the console wallet, but must be tested with the FFI
in Aurora.

What process can a PR reviewer use to test or verify this change?
---
Code review.

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->