feat: improve mempool error msg when mempool out of sync (#6618)
Description
---
Improved mempool error message when mempool falls behind the persisted
blockchain db.

Motivation and Context
---
This error will occur in an edge case when the mempool is in the process
of finalizing a block template and just before it is done, a new block
has been added to the blockchain db. As the two processes are
asynchronous, the resolution implemented in the code is to return an
error message so that the client can issue a new request. Meanwhile, the
mempool has time to react to the updated blockchain db state.

How Has This Been Tested?
---
Compile succeed.

What process can a PR reviewer use to test or verify this change?
---
Code review.

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->