feat: add checked_sub (#6628)
Description
---
Add checked_sub implementation for accumulated difficulty struct. 

Motivation and Context
---
In a normal block chain world, you would never subtract difficulty, but
in P2pool you only count the last X blocks. This means that when a block
falls away and is not used anymore, we need to subtract its difficulty.