test: add merge mining proxy monerod stress test code (#6639)
Description
---
Add merge mining proxy monerod stress test code used to perform stress
tests for #6637.

Motivation and Context
---
The tests may need to be performed again in the future.

How Has This Been Tested?
---
Tested as part of #6637.

What process can a PR reviewer use to test or verify this change?
---
Code review.

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->