Add check connections
Added check connections to the base nodes application. All active connections are pinged
on a set (slowish) interval and the nodes that do not respond timeousely on three consecutive
iterations with a corresponding pong are disconnected. This will help to keep the list of
active connections (lazily) up to date.