feat: trickle validators (#6660)
Description
---
We allow here only a configured (in Consensus constants) number of
validator nodes to be active in an epoch, so this way we avoid having
way too much validators appear at the same time.
All the validator nodes are saved to the blockchain database, when there
is a new validator node registration UTXO, but with a different `start
epoch` that is calculated from the already saved validator nodes.
There is a new gRPC method called `GetValidatorNodeChanges`, where we
can ask base node to give back all the validator node additions and
deletions in the given start/end height range.
When Layer-2 scans new blocks and sees that there is a change in the
validator node set, simply calls this new grpc method and adds/removes
validator nodes. This way base node controls fully how many nodes are
registered at which epoch.

Motivation and Context
---
Layer-2 validator nodes are registered all at once when they were added
to the chain with a special UTXO, but it allows to register too many
nodes at the same time, so consensus can catch up much slower.

How Has This Been Tested?
---

What process can a PR reviewer use to test or verify this change?
---

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->

---------

Co-authored-by: Stan Bondi <sdbondi@users.noreply.github.com>