test!: fix cucumber on libp2p feature branch (#6692)
Description
---
FIx cucumber tests on libp2p feature branch

_**Notes:**_
- _Instabilities with cucumber when using concurrency other than
`--concurrency 1` were detected locally, more so for the wallet FFI
suite of tests than the normal tests, but `--concurrency 1` is
recommended for both. This is because addresses are mixed (discoverable)
with libp2p when running in the single cucumber process._
- _Outstanding issue #6693 to be addressed in a consecutive PR._

Motivation and Context
---
Some tests were failing

How Has This Been Tested?
---
- `cargo test --release --test cucumber -- --tags "@wallet-ffi and not
@broken" --concurrency 1 --retry 2`
- `cargo test --release --all-features --test cucumber -- --tags
"@critical and (not @long-running) and (not @wallet-ffi) and (not
@chat-ffi) and (not @broken)" --concurrency 1 --retry 2`

What process can a PR reviewer use to test or verify this change?
---
Code review
Run the cucumber tests

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
BREAKING CHANGE: Wallet FFI interface change on `pub unsafe extern "C"
fn comms_config_create`

---------

Co-authored-by: SW van Heerden <swvheerden@gmail.com>