Fix base node shutdown
Fixed the base node shutdown on Ubuntu whereas pressing Ctrl-C multiple times rendered the
command line interface in an undefined state. This was not issue on Windows.

Note: The change to 'base_layer\core\src\base_node\tari_pulse_service\mod.rs' is just to
enable testing (get rid of the repeated command-line spam), and will be fixed in #6696.