Fix mempool sync start early
Mempool sync can only start when `is_synced` is set to `true`.
- An edge case was witnessed whereby mempool sync can start early if the entire network is generally
slow to respond to pings. The `ChainMetadataEvent::NetworkSilence` event is transmitted if 3x
consecutive ping cycles were started without corresponding pongs received, which sets
`is_synced` to `true`. The moment that peer metadata is then received, block sync and mempool sync
will occur at the same time. For the time being the consecutive ping cycle threshold was increased
to 4x and a warning message was added to let the user know this may be the case.
- Additional debug logs were added to track all events that can set the `is_synced` to `true`.
- Removed 'log-mdc' form the dependencies as the relevant code is not used