Fix seedstrap seeds disconnect
Fixed seedstrap seeds disconnecting during seedstrap phase. This was caused by the
health service that obtained conenctions to all seed peers and disconnected them
while seedstrap was busy. This was fixed by adding a soft disconnect alongside the
hard disconnect in `peer_connection.rs::PeerConnection` so that parallel spawned
process can politely request connections to be broken if not in use. Also added
better logging information for peer disconnects in the the requester is also logged.