Prevent peers syncing local addresses
This PR prevents peers sending local addresses during peer sync when peer
info is requested. The rationale here is that peer sync is done by a local
node to learn about the network and then for the local node to try and
contact those peers. It is not possible for a local node to contact a remote
peer on its internal address.