Add tolerance for invalid peer info
Added tolerance for banable offences while busy with peer sync. A sync session will be
interrupted after a configurable number of banable offencses has been committed. When the
sync session ends due to this thresholf or otherwise, the peer will be banned for the most
severe ban offence. This scheme enables a local node to accumulate good peer info sets in
an optimal manner.