Add seed exclusion to sync
Force header and block sync to use non-seed peers. The rationale here is that the seed peers'
function is primarily to provide other nodes with network peer information when they have no
peer database, typically when starting a new node or when the peer database is manually deleted.
Also, no miners are/should be connected to seed nodes. As such, seed nodes will always be
followers in the blockchain network and not leaders.