Improve header and block sync
This PR changes the sync strategy from an “all headers → all blocks” approach
to an incremental session-based process. This is done by defining a session
target number of headers to sync, followed by the same number of blocks to sync,
then repeated until we reached the sync peer's advertized block tip. This way,
whenever header and/or block sync is interrupted, we would have "banked" the
good sync work we have done.