Add actual fee to memo field
Added actual fee to the memo field by calculating it in a circular manner. We
need to the actual fee before we can create the memo field, but the memo field
is used in the fee calculation, so we repeat non-persistent steps to get the fee.

This fee-before-memo calculation was added to all pertinent places in the
transaction service.